Add methods for streaming (de)compression of direct ByteBuffers.
The goal of this change is to enable the highest-performance (de)compression of streams. `ZSTD_compressStream2` and `ZSTD_decompressStream` are exposed on `ZstdCompressCtx` and `ZstdDecompressCtx` respectively. They work in terms of direct `ByteBuffer`s and allow non-blocking, incremental handling of streams. The new `reset()` methods on `ZstdCompressCtx` and `ZstdDecompressCtx` allow reusing contexts, avoiding unnecessary allocations and deallocations.